refactor: address review - entrypoints/dashboard.py, skill structure
The launcher-side helper and the console script move to
entrypoints/dashboard.py (main delegates to the server, matching the
other entrypoints); the CACHE_DIR docstring drops the verifiers
mention; the dashboard skill gets a dedicated Isolated mode section
and a 'Dependency-free dashboard' heading.
