+++ src/prime_rl/entrypoints/evals.py
@@ -19,7 +19,7 @@ def main():
 
     write_launch_toml(config.output_dir, "evals")
     if config.dashboard:
-        from prime_rl.utils.dashboard_daemon import ensure_dashboard
+        from prime_rl.entrypoints.dashboard import ensure_dashboard
         from prime_rl.utils.logger import get_logger, setup_logger
 
         ensure_dashboard(config.output_dir, get_logger() or setup_logger(config.log.level))
